requirements.txt
bcrypt==3.2.2
Flask==3.0.2
Flask-Login==0.6.3
mysql-connector-python==8.0.15
python-dotenv==1.0.1
